add reset period for translate and punch in one moment
